bot/requirements.txt
python-telegram-bot>=21,<22

asyncpg>=0.29,<1
